685/PJ2/laptop_requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.74.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
pip==23.0.1
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
setuptools==66.1.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
paho-mqtt==1.5.0
